requirements.txt
absl-py==1.1.0
accelerate==0.11.0
acvl_utils==0.2.5
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
babel==2.16.0
batchgenerators==0.25.1
batchgeneratorsv2==0.2.3
beautifulsoup4==4.12.3
bleach==6.2.0
blessed==1.20.0
blobfile==2.1.1
blosc2==3.0.0b4
cachetools==5.2.0
certifi==2022.6.15
cffi==1.17.1
charset-normalizer==2.0.12
click==8.1.3
comm==0.2.2
connected-components-3d==3.23.0
contextlib2==21.6.0
contourpy==1.3.1
cycler==0.11.0
dacite==1.9.1
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
deprecation==2.1.0
dicom2nifti==2.5.1
docker-pycreds==0.4.0
dynamic_network_architectures==0.3.1
einops==0.4.1
einops-exts==0.0.3
elasticdeform==0.5.1
ema-pytorch==0.0.8
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
fft-conv-pytorch==1.2.0
filelock==3.15.4
flatbuffers==25.2.10
fonttools==4.34.4
fqdn==1.5.1
frozenlist==1.3.0
fsspec==2022.5.0
ftfy==6.1.1
future==0.18.2
gast==0.6.0
gitdb==4.0.9
GitPython==3.1.27
glob2==0.7
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpustat==1.1.1
graphviz==0.20.3
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
hf-deepali==0.6.3
hiddenlayer @ git+https://github.com/FabianIsensee/hiddenlayer.git@b7263b6dc4569da1b6dea5964e1eac78fa32fa77
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.17.3
humanize==4.2.2
hydra-core==1.2.0
idna==3.3
imagecodecs==2024.12.30
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel==6.29.5
ipython==8.31.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.1.0
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.9.0
kiwisolver==1.4.3
libclang==18.1.1
linecache2==1.0.0
lxml==4.9.1
Markdown==3.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.2
matplotlib-inline==0.1.7
mdurl==0.1.2
MedPy==0.3.0
mistune==3.1.0
ml_collections==0.1.1
ml_dtypes==0.5.1
monai==0.9.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.2
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
ndindex==1.9.2
nest-asyncio==1.6.0
networkx==2.8.5
nibabel==4.0.1
nilearn==0.9.1
-e git+https://github.com/MIC-DKFZ/nnUNet.git@f8f5b494b7226b8b0b1bca34ad3e9b68facb52b8#egg=nnunetv2
notebook==7.3.2
notebook_shim==0.2.4
numexpr==2.10.2
numpy==1.23.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
omegaconf==2.2.3
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.14.1
overrides==7.7.0
packaging==24.1
pandas==1.4.3
pandocfilters==1.5.1
parso==0.8.4
pathtools==0.1.2
pillow==11.1.0
platformdirs==4.3.6
prometheus_client==0.21.1
promise==2.3
prompt_toolkit==3.0.48
protobuf==3.20.1
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.22
pycryptodomex==3.21.0
pyDeprecate==0.3.1
pydicom==2.3.0
Pygments==2.19.1
PYMIC==0.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-gdcm==3.0.24.1
python-json-logger==3.2.1
pytorch-lightning==1.6.4
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==19.0.2
referencing==0.35.1
regex==2022.6.2
requests==2.28.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rotary-embedding-torch==0.1.5
rpds-py==0.22.3
rsa==4.8
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.4.4
scikit-image==0.19.3
scikit-learn==1.1.2
scikit-video==1.1.11
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.3
sentry-sdk==1.7.2
setproctitle==1.2.3
shortuuid==1.0.9
SimpleITK==2.4.1
six==1.16.0
sk-video==1.1.10
smmap==5.0.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.1.0
tifffile==2022.8.3
timm==1.0.8
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
torch==2.1.2
torch-tb-profiler==0.4.0
torchio==0.18.80
torchmetrics==0.9.1
torchstat==0.0.7
torchvision==0.16.2
tornado==6.4.2
tqdm==4.64.0
traceback2==1.4.0
traitlets==5.14.3
triton==2.1.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
unittest2==1.1.0
UNKNOWN @ git+https://github.com/BioMedIA/deepali.git@ae3e4762a20cbbd0711cc5dd53a7f1aafad49b86
uri-template==1.3.0
urllib3==1.26.9
vedo==2025.5.3
vtk==9.4.1
wandb==0.12.21
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.1.2
wrapt==1.14.1
yacs==0.1.8
yarl==1.7.2
zipp==3.8.0
